completed search user

+++ src/views/ProfileTeacher.php
@@ -82,15 +82,17 @@
     <main>
         <!-- top section containing back button, username_ID and edit button -->
         <div class="response">
-            <a href="mainpageTeacher.php"><button class="flex_button"><span class="material-symbols-outlined">arrow_back_ios</span>Go Back</button></a>
-            <h1>Teacher profile</h1>
-            <?php
-                if($_SESSION['sourcepage'] == "searchUser"){
-                    echo <<<HTML
-                            <a href="deleteProfile.php?id='$_SESSION[delete_id]'"><button class="flex_button"><span class="material-symbols-outlined">edit</span>Delete Profile</button></a>
-                HTML;}else{ echo <<<HTML
-                    <a href="editTeacher.php"><button class="flex_button"><span class="material-symbols-outlined">edit</span>Edit some information</button></a>
-                HTML;}?>
+        <?php
+            if(isset($_SESSION['sourcepage']) && $_SESSION['sourcepage'] == "searchUser"){
+                echo <<<HTML
+                    <a href="searchUser.php"><button class="flex_button"><span class="material-symbols-outlined">arrow_back_ios</span>Go Back</button></a>
+                    <h1>Teacher profile</h1>
+                    <a href="deleteProfile.php?id='$_SESSION[delete_id]'"><button class="flex_button"><span class="material-symbols-outlined">edit</span>Delete Profile</button></a>
+            HTML;unset($_SESSION['sourcepage']);}else{ echo <<<HTML
+                <a href="mainpageTeacher.php"><button class="flex_button"><span class="material-symbols-outlined">arrow_back_ios</span>Go Back</button></a>
+                <h1>Teacher profile</h1>
+                <a href="editTeacher.php"><button class="flex_button"><span class="material-symbols-outlined">edit</span>Edit some information</button></a>
+            HTML;}?>
         </div>
 
         <!-- middle section containing user info, teacher into -->
